backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
yara-python